// src/app/services/api/models/yahoo/yahoo-api-get-statistics-result.ts
export interface YahooApiGetStatisticsResultAnnualHoldingsTurnover {
}

export interface YahooApiGetStatisticsResultEnterpriseToRevenue {
    raw: number;
    fmt: string;
}

export interface YahooApiGetStatisticsResultBeta3Year {
}

export interface YahooApiGetStatisticsResultProfitMargins {
    raw: number;
    fmt: string;
}

export interface YahooApiGetStatisticsResultEnterpriseToEbitda {
    raw: number;
    fmt: string;
}

export interface YahooApiGetStatisticsResult52WeekChange {
    raw: number;
    fmt: string;
}

export interface YahooApiGetStatisticsResultMorningStarRiskRating {
}

export interface YahooApiGetStatisticsResultForwardEps {
    raw: number;
    fmt: string;
}

export interface YahooApiGetStatisticsResultRevenueQuarterlyGrowth {
}

export interface YahooApiGetStatisticsResultSharesOutstanding {
    raw: number;
    fmt: string;
    longFmt: string;
}

export interface YahooApiGetStatisticsResultFundInceptionDate {
}
// ...
export interface YahooApiGetStatisticsResultDefaultKeyStatistics {
    annualHoldingsTurnover: YahooApiGetStatisticsResultAnnualHoldingsTurnover;
    enterpriseToRevenue: YahooApiGetStatisticsResultEnterpriseToRevenue;
    beta3Year: YahooApiGetStatisticsResultBeta3Year;
    profitMargins: YahooApiGetStatisticsResultProfitMargins;
    enterpriseToEbitda: YahooApiGetStatisticsResultEnterpriseToEbitda;
    _52WeekChange: YahooApiGetStatisticsResult52WeekChange;
    morningStarRiskRating: YahooApiGetStatisticsResultMorningStarRiskRating;
    forwardEps: YahooApiGetStatisticsResultForwardEps;
    revenueQuarterlyGrowth: YahooApiGetStatisticsResultRevenueQuarterlyGrowth;
    sharesOutstanding: YahooApiGetStatisticsResultSharesOutstanding;
    fundInceptionDate: YahooApiGetStatisticsResultFundInceptionDate;
    annualReportExpenseRatio: YahooApiGetStatisticsResultAnnualReportExpenseRatio;
    totalAssets: YahooApiGetStatisticsResultTotalAssets;
    bookValue: YahooApiGetStatisticsResultBookValue;
    sharesShort: YahooApiGetStatisticsResultSharesShort;
    sharesPercentSharesOut: YahooApiGetStatisticsResultSharesPercentSharesOut;
    fundFamily?: any;
    lastFiscalYearEnd: YahooApiGetStatisticsResultLastFiscalYearEnd;
    heldPercentInstitutions: YahooApiGetStatisticsResultHeldPercentInstitutions;
    netIncomeToCommon: YahooApiGetStatisticsResultNetIncomeToCommon;
    trailingEps: YahooApiGetStatisticsResultTrailingEps;
    lastDividendValue: YahooApiGetStatisticsResultLastDividendValue;
    SandP52WeekChange: YahooApiGetStatisticsResultSandP52WeekChange;
    priceToBook: YahooApiGetStatisticsResultPriceToBook;
    heldPercentInsiders: YahooApiGetStatisticsResultHeldPercentInsiders;
    nextFiscalYearEnd: YahooApiGetStatisticsResultNextFiscalYearEnd;
    yield: YahooApiGetStatisticsResultYield;
    mostRecentQuarter: YahooApiGetStatisticsResultMostRecentQuarter;
    shortRatio: YahooApiGetStatisticsResultShortRatio;
    sharesShortPreviousMonthDate: YahooApiGetStatisticsResultSharesShortPreviousMonthDate;
    floatShares: YahooApiGetStatisticsResultFloatShares;
    beta: YahooApiGetStatisticsResultBeta;
    enterpriseValue: YahooApiGetStatisticsResultEnterpriseValue;
    priceHint: YahooApiGetStatisticsResultPriceHint;
    threeYearAverageReturn: YahooApiGetStatisticsResultThreeYearAverageReturn;
    lastSplitDate: YahooApiGetStatisticsResultLastSplitDate;
    lastSplitFactor: string;
    legalType?: any;
    lastDividendDate: YahooApiGetStatisticsResultLastDividendDate;
    morningStarOverallRating: YahooApiGetStatisticsResultMorningStarOverallRating;
    earningsQuarterlyGrowth: YahooApiGetStatisticsResultEarningsQuarterlyGrowth;
    priceToSalesTrailing12Months: YahooApiGetStatisticsResultPriceToSalesTrailing12Months;
    dateShortInterest: YahooApiGetStatisticsResultDateShortInterest;
    pegRatio: YahooApiGetStatisticsResultPegRatio;
    ytdReturn: YahooApiGetStatisticsResultYtdReturn;
    forwardPE: YahooApiGetStatisticsResultForwardPE;
    maxAge: number;
    lastCapGain: YahooApiGetStatisticsResultLastCapGain;
    shortPercentOfFloat: YahooApiGetStatisticsResultShortPercentOfFloat;
    sharesShortPriorMonth: YahooApiGetStatisticsResultSharesShortPriorMonth;
    impliedSharesOutstanding: YahooApiGetStatisticsResultImpliedSharesOutstanding;
    category?: any;
    fiveYearAverageReturn: YahooApiGetStatisticsResultFiveYearAverageReturn;
}

export interface YahooApiGetStatisticsResultFinancialsTemplate {
    code: string;
    maxAge: number;
}
// ...
export interface YahooApiGetStatisticsResultPrice {
    quoteSourceName: string;
    regularMarketOpen: YahooApiGetStatisticsResultRegularMarketOpen;
    averageDailyVolume3Month: YahooApiGetStatisticsResultAverageDailyVolume3Month;
    exchange: string;
    regularMarketTime: number;
    volume24Hr: YahooApiGetStatisticsResultVolume24Hr;
    regularMarketDayHigh: YahooApiGetStatisticsResultRegularMarketDayHigh;
    shortName: string;
    averageDailyVolume10Day: YahooApiGetStatisticsResultAverageDailyVolume10Day;
    longName: string;
    regularMarketChange: YahooApiGetStatisticsResultRegularMarketChange;
    currencySymbol: string;
    regularMarketPreviousClose: YahooApiGetStatisticsResultRegularMarketPreviousClose;
    postMarketTime: number;
    preMarketPrice: YahooApiGetStatisticsResultPreMarketPrice;
    preMarketTime: number;
    exchangeDataDelayedBy: number;
    toCurrency?: any;
    postMarketChange: YahooApiGetStatisticsResultPostMarketChange;
    postMarketPrice: YahooApiGetStatisticsResultPostMarketPrice;
    exchangeName: string;
    preMarketChange: YahooApiGetStatisticsResultPreMarketChange;
    circulatingSupply: YahooApiGetStatisticsResultCirculatingSupply;
    regularMarketDayLow: YahooApiGetStatisticsResultRegularMarketDayLow;
    priceHint: YahooApiGetStatisticsResultPriceHint2;
    currency: string;
    regularMarketPrice: YahooApiGetStatisticsResultRegularMarketPrice;
    regularMarketVolume: YahooApiGetStatisticsResultRegularMarketVolume;
    lastMarket?: any;
    regularMarketSource: string;
    openInterest: YahooApiGetStatisticsResultOpenInterest;
    marketState: string;
    underlyingSymbol?: any;
    marketCap: YahooApiGetStatisticsResultMarketCap;
    quoteType: string;
    preMarketChangePercent: YahooApiGetStatisticsResultPreMarketChangePercent;
    volumeAllCurrencies: YahooApiGetStatisticsResultVolumeAllCurrencies;
    postMarketSource: string;
    strikePrice: YahooApiGetStatisticsResultStrikePrice;
    symbol: string;
    postMarketChangePercent: YahooApiGetStatisticsResultPostMarketChangePercent;
    preMarketSource: string;
    maxAge: number;
    fromCurrency?: any;
    regularMarketChangePercent: YahooApiGetStatisticsResultRegularMarketChangePercent;
}
// ...
export interface YahooApiGetStatisticsResultFinancialData {
    ebitdaMargins: YahooApiGetStatisticsResultEbitdaMargins;
    profitMargins: YahooApiGetStatisticsResultProfitMargins2;
    grossMargins: YahooApiGetStatisticsResultGrossMargins;
    operatingCashflow: YahooApiGetStatisticsResultOperatingCashflow;
    revenueGrowth: YahooApiGetStatisticsResultRevenueGrowth;
    operatingMargins: YahooApiGetStatisticsResultOperatingMargins;
    ebitda: YahooApiGetStatisticsResultEbitda;
    targetLowPrice: YahooApiGetStatisticsResultTargetLowPrice;
    recommendationKey: string;
    grossProfits: YahooApiGetStatisticsResultGrossProfits;
    freeCashflow: YahooApiGetStatisticsResultFreeCashflow;
    targetMedianPrice: YahooApiGetStatisticsResultTargetMedianPrice;
    currentPrice: YahooApiGetStatisticsResultCurrentPrice;
    earningsGrowth: YahooApiGetStatisticsResultEarningsGrowth;
    currentRatio: YahooApiGetStatisticsResultCurrentRatio;
    returnOnAssets: YahooApiGetStatisticsResultReturnOnAssets;
    numberOfAnalystOpinions: YahooApiGetStatisticsResultNumberOfAnalystOpinions;
    targetMeanPrice: YahooApiGetStatisticsResultTargetMeanPrice;
    debtToEquity: YahooApiGetStatisticsResultDebtToEquity;
    returnOnEquity: YahooApiGetStatisticsResultReturnOnEquity;
    targetHighPrice: YahooApiGetStatisticsResultTargetHighPrice;
    totalCash: YahooApiGetStatisticsResultTotalCash;
    totalDebt: YahooApiGetStatisticsResultTotalDebt;
    totalRevenue: YahooApiGetStatisticsResultTotalRevenue;
    totalCashPerShare: YahooApiGetStatisticsResultTotalCashPerShare;
    financialCurrency: string;
    maxAge: number;
    revenuePerShare: YahooApiGetStatisticsResultRevenuePerShare;
    quickRatio: YahooApiGetStatisticsResultQuickRatio;
    recommendationMean: YahooApiGetStatisticsResultRecommendationMean;
}

export interface YahooApiGetStatisticsResultQuoteType {
    exchange: string;
    shortName: string;
    longName: string;
    exchangeTimezoneName: string;
    exchangeTimezoneShortName: string;
    isEsgPopulated: boolean;
    gmtOffSetMilliseconds: string;
    quoteType: string;
    symbol: string;
    messageBoardId: string;
    market: string;
}
// ...
export interface YahooApiGetStatisticsResultCalendarEvents {
    maxAge: number;
    earnings: YahooApiGetStatisticsResultEarnings;
    exDividendDate: YahooApiGetStatisticsResultExDividendDate;
    dividendDate: YahooApiGetStatisticsResultDividendDate;
}
// ...
export interface YahooApiGetStatisticsResultPageViews {
    shortTermTrend: string;
    midTermTrend: string;
    longTermTrend: string;
    maxAge: number;
}
// ...
export interface YahooApiGetStatisticsResultQuoteData {
    AAPL: YahooApiGetStatisticsResultAAPL;
    AMZN: YahooApiGetStatisticsResultAMZN;
    TSLA: YahooApiGetStatisticsResultTSLA;
    FB: YahooApiGetStatisticsResultFB;
    GOOG: YahooApiGetStatisticsResultGOOG;
    NFLX: YahooApiGetStatisticsResultNFLX;
}

export interface YahooApiGetStatisticsResultMktmData {
}
// ...
export interface YahooApiGetStatisticsResultTimeSeries {
    quarterlyPeRatio: YahooApiGetStatisticsResultQuarterlyPeRatio[];
    quarterlyForwardPeRatio: YahooApiGetStatisticsResultQuarterlyForwardPeRatio[];
    trailingEnterprisesValueEBITDARatio: YahooApiGetStatisticsResultTrailingEnterprisesValueEBITDARatio[];
    quarterlyPegRatio: YahooApiGetStatisticsResultQuarterlyPegRatio[];
    trailingForwardPeRatio: YahooApiGetStatisticsResultTrailingForwardPeRatio[];
    trailingMarketCap: YahooApiGetStatisticsResultTrailingMarketCap[];
    quarterlyPsRatio: YahooApiGetStatisticsResultQuarterlyPsRatio[];
    quarterlyMarketCap: YahooApiGetStatisticsResultQuarterlyMarketCap[];
    trailingEnterpriseValue: YahooApiGetStatisticsResultTrailingEnterpriseValue[];
    quarterlyEnterprisesValueRevenueRatio: YahooApiGetStatisticsResultQuarterlyEnterprisesValueRevenueRatio[];
    quarterlyEnterprisesValueEBITDARatio: YahooApiGetStatisticsResultQuarterlyEnterprisesValueEBITDARatio[];
    trailingPegRatio: YahooApiGetStatisticsResultTrailingPegRatio[];
    trailingPeRatio: YahooApiGetStatisticsResultTrailingPeRatio[];
    quarterlyEnterpriseValue: YahooApiGetStatisticsResultQuarterlyEnterpriseValue[];
    trailingEnterprisesValueRevenueRatio: YahooApiGetStatisticsResultTrailingEnterprisesValueRevenueRatio[];
    quarterlyPbRatio: YahooApiGetStatisticsResultQuarterlyPbRatio[];
    trailingPbRatio: YahooApiGetStatisticsResultTrailingPbRatio[];
    trailingPsRatio: YahooApiGetStatisticsResultTrailingPsRatio[];
    timestamp: number[];
}

export interface YahooApiGetStatisticsResultMeta {
    symbol: string;
    start: number;
    end: number;
    timeUnit: string;
}

export interface YahooApiGetStatisticsResult {
    defaultKeyStatistics: YahooApiGetStatisticsResultDefaultKeyStatistics;
    financialsTemplate: YahooApiGetStatisticsResultFinancialsTemplate;
    price: YahooApiGetStatisticsResultPrice;
    financialData: YahooApiGetStatisticsResultFinancialData;
    quoteType: YahooApiGetStatisticsResultQuoteType;
    calendarEvents: YahooApiGetStatisticsResultCalendarEvents;
    summaryDetail: YahooApiGetStatisticsResultSummaryDetail;
    symbol: string;
    pageViews: YahooApiGetStatisticsResultPageViews;
    quoteData: YahooApiGetStatisticsResultQuoteData;
    mktmData: YahooApiGetStatisticsResultMktmData;
    timeSeries: YahooApiGetStatisticsResultTimeSeries;
    meta: YahooApiGetStatisticsResultMeta;
    loading: boolean;
    errorList: any[];
}